requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
sklearn
statistics
dill
xgboost
flask
cassandra-driver
touch
#-e .